rename cosmos from "migration" to "cosmos" simply
