+++ src/Orleans.Core.Abstractions/Properties/AssemblyInfo.cs
@@ -8,14 +8,14 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("DefaultCluster.Tests")]
 [assembly: InternalsVisibleTo("NonSilo.Tests")]
 [assembly: InternalsVisibleTo("Tester.AzureUtils")]
-[assembly: InternalsVisibleTo("Tester.Reminders.Cosmos.Migration")]
+[assembly: InternalsVisibleTo("Tester.Reminders.Cosmos")]
 [assembly: InternalsVisibleTo("AWSUtils.Tests")]
 [assembly: InternalsVisibleTo("TesterInternal")]
 [assembly: InternalsVisibleTo("TestInternalGrainInterfaces")]
 [assembly: InternalsVisibleTo("TestInternalGrains")]
 
 [assembly: InternalsVisibleTo("Orleans.Persistence.Migration")]
-[assembly: InternalsVisibleTo("Orleans.Reminders.Cosmos.Migration")]
+[assembly: InternalsVisibleTo("Orleans.Reminders.Cosmos")]
 [assembly: InternalsVisibleTo("Tester.AzureUtils.Migration")]
 
 // Mocking libraries
