+++ src/Orleans.Persistence.Migration/Orleans.Persistence.Migration.csproj
@@ -29,7 +29,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <InternalsVisibleTo Include="Tester.Reminders.Cosmos.Migration"/>
+    <InternalsVisibleTo Include="Tester.Reminders.Cosmos"/>
   </ItemGroup>
 
 </Project>
