+++ test/Extensions/Tester.Persistence.Cosmos/Tester.Persistence.Cosmos.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Azure\Orleans.Reminders.Cosmos.Migration\Orleans.Reminders.Cosmos.Migration.csproj" />
+    <ProjectReference Include="..\..\..\src\Azure\Orleans.Reminders.Cosmos\Orleans.Reminders.Cosmos.csproj" />
     <ProjectReference Include="..\..\..\src\Orleans.Persistence.Migration\Orleans.Persistence.Migration.csproj" />
     <ProjectReference Include="..\..\TesterInternal\TesterInternal.csproj" />
     <ProjectReference Include="..\Tester.AzureUtils.Migration\Tester.AzureUtils.Migration.csproj" />
