+++ test/Extensions/Tester.Reminders.Cosmos.Migration/Setup/RemindersCosmosSetup.cs
@@ -1,5 +1,4 @@
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Options;
 using Orleans.Hosting;
 using Orleans.Persistence.AzureStorage.Migration;
 using Orleans.Persistence.AzureStorage.Migration.Reminders;
@@ -7,13 +6,13 @@ using Orleans.Persistence.Cosmos;
 using Orleans.Persistence.Cosmos.DocumentIdProviders;
 using Orleans.Persistence.Migration;
 using Orleans.TestingHost;
-using Tester.Reminders.Cosmos.Migration.Helpers;
-using Tester.Reminders.Cosmos.Migration.Tests;
 using Xunit;
 using TesterInternal.AzureInfra;
+using Tester.Reminders.Cosmos.Tests;
+using Tester.Reminders.Cosmos.Helpers;
 using Resources = Tester.AzureUtils.Migration.Resources;
 
-namespace Tester.Reminders.Cosmos.Migration.Setup
+namespace Tester.Reminders.Cosmos.Setup
 {
     [TestCategory("Functional"), TestCategory("Migration"), TestCategory("Azure"), TestCategory("Reminders"), TestCategory("Cosmos")]
     public class RemindersCosmosSetup : RemindersCosmosTests, IClassFixture<RemindersCosmosSetup.Fixture>
